requirements.txt
-i https://pypi.org/simple
attrs==23.1.0; python_version >= '3.7'
blinker==1.6.3; python_version >= '3.7'
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.3.0; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
coverage[toml]==7.3.2; python_version >= '3.8'
exceptiongroup==1.1.3
flask==3.0.0
flask-sqlalchemy==3.1.1
idna==3.4; python_version >= '3.5'
iniconfig==2.0.0; python_version >= '3.7'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
markupsafe==2.1.3; python_version >= '3.7'
marshmallow==3.20.1
packaging==23.2; python_version >= '3.7'
pluggy==1.3.0; python_version >= '3.8'
psycopg2-binary==2.9.9
pytest==7.2.2
pytest-cov==4.0.0
pytest-env==0.8.1
pytest-mock==3.10.0
python-dotenv==1.0.0
requests==2.31.0
sqlalchemy==2.0.21
typing-extensions==4.8.0; python_version >= '3.8'
urllib3==2.0.6; python_version >= '3.7'
